refactor: isolate styling primitives for non-windows
